Add result correction feature
Implemented a feature that allows the user to correct the displayed
result.
Previously, pressing a digit key after getting a result would reset
the calculation.
Now, if the user presses the delete button and then enters a digit key,
the `number1` value will be corrected without resetting
the entire calculation.